requirements.txt
fastapi
uvicorn
pillow
numpy
scipy
python-multipart